+++ src/main/java/gov/nist/csd/pm/pap/AccessQuerier.java
@@ -1,8 +1,9 @@
-package gov.nist.csd.pm.pap.query;
+package gov.nist.csd.pm.pap;
 
+import gov.nist.csd.pm.pap.query.AccessQuery;
 import gov.nist.csd.pm.pap.store.PolicyStore;
 
-public abstract class AccessQuerier extends Querier implements AccessQuery{
+public abstract class AccessQuerier extends Querier implements AccessQuery {
 
     public AccessQuerier(PolicyStore store) {
         super(store);
