+++ src/main/java/gov/nist/csd/pm/pap/PAP.java
@@ -4,26 +4,25 @@ import gov.nist.csd.pm.pap.admin.AdminPolicy;
 import gov.nist.csd.pm.pap.executable.AdminExecutable;
 import gov.nist.csd.pm.pap.executable.AdminExecutor;
 import gov.nist.csd.pm.pap.modification.PolicyModification;
-import gov.nist.csd.pm.pap.modification.PolicyModifier;
 import gov.nist.csd.pm.pap.pml.PMLCompiler;
 import gov.nist.csd.pm.pap.pml.context.ExecutionContext;
 import gov.nist.csd.pm.pap.pml.executable.operation.PMLOperation;
 import gov.nist.csd.pm.pap.pml.executable.routine.PMLRoutine;
 import gov.nist.csd.pm.pap.pml.statement.PMLStatement;
 import gov.nist.csd.pm.pap.pml.value.Value;
-import gov.nist.csd.pm.pap.query.PolicyQuerier;
 import gov.nist.csd.pm.pap.query.PolicyQuery;
+import gov.nist.csd.pm.pap.query.UserContext;
 import gov.nist.csd.pm.pap.serialization.PolicyDeserializer;
 import gov.nist.csd.pm.pap.serialization.PolicySerializer;
 import gov.nist.csd.pm.pap.exception.PMException;
-import gov.nist.csd.pm.pap.query.UserContext;
 import gov.nist.csd.pm.pap.store.PolicyStore;
+import gov.nist.csd.pm.pap.tx.Transactional;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public abstract class PAP implements PolicyPoint, AdminExecutor {
+public abstract class PAP implements AdminExecutor, Transactional {
 
     protected final PolicyStore policyStore;
     private final PolicyModifier modifier;
@@ -50,15 +49,12 @@ public abstract class PAP implements PolicyPoint, AdminExecutor {
         return policyStore;
     }
 
-    @Override
     public abstract PolicyQuery query();
 
-    @Override
     public PolicyModification modify() {
         return modifier;
     }
 
-    @Override
     public void reset() throws PMException {
         policyStore.reset();
