+++ src/main/java/gov/nist/csd/pm/pap/PolicyQuerier.java
@@ -1,5 +1,6 @@
-package gov.nist.csd.pm.pap.query;
+package gov.nist.csd.pm.pap;
 
+import gov.nist.csd.pm.pap.query.PolicyQuery;
 import gov.nist.csd.pm.pap.store.PolicyStore;
 
 public abstract class PolicyQuerier extends Querier implements PolicyQuery {
