+++ src/main/java/gov/nist/csd/pm/pap/Querier.java
@@ -1,4 +1,4 @@
-package gov.nist.csd.pm.pap.query;
+package gov.nist.csd.pm.pap;
 
 import gov.nist.csd.pm.pap.store.PolicyStore;
