+++ src/main/java/gov/nist/csd/pm/pap/RoutinesModifier.java
@@ -1,11 +1,12 @@
-package gov.nist.csd.pm.pap.modification;
+package gov.nist.csd.pm.pap;
 
 import gov.nist.csd.pm.pap.exception.PMException;
 import gov.nist.csd.pm.pap.exception.RoutineExistsException;
+import gov.nist.csd.pm.pap.modification.RoutinesModification;
 import gov.nist.csd.pm.pap.routine.Routine;
 import gov.nist.csd.pm.pap.store.PolicyStore;
 
-public class RoutinesModifier extends Modifier implements RoutinesModification{
+public class RoutinesModifier extends Modifier implements RoutinesModification {
 
     public RoutinesModifier(PolicyStore store) {
         super(store);
