+++ src/main/java/gov/nist/csd/pm/pap/query/explain/ExplainNode.java
@@ -1,7 +1,6 @@
 package gov.nist.csd.pm.pap.query.explain;
 
 import java.util.List;
-import java.util.Objects;
 
 public record ExplainNode(String node, List<ExplainAssociation> associations) {
