+++ src/main/java/gov/nist/csd/pm/pdp/modification/ObligationsModificationAdjudicator.java
@@ -6,7 +6,6 @@ import gov.nist.csd.pm.pap.obligation.Obligation;
 import gov.nist.csd.pm.epp.EventEmitter;
 import gov.nist.csd.pm.pap.PAP;
 import gov.nist.csd.pm.pap.exception.PMException;
-import gov.nist.csd.pm.pap.modification.ObligationsModifier;
 import gov.nist.csd.pm.pap.op.PrivilegeChecker;
 import gov.nist.csd.pm.pap.op.obligation.CreateObligationOp;
 import gov.nist.csd.pm.pap.op.obligation.DeleteObligationOp;
