+++ src/main/java/gov/nist/csd/pm/pdp/query/PolicyQueryAdjudicator.java
@@ -1,6 +1,7 @@
 package gov.nist.csd.pm.pdp.query;
 
 import gov.nist.csd.pm.pap.PAP;
+import gov.nist.csd.pm.pap.query.UserContext;
 import gov.nist.csd.pm.pap.op.PrivilegeChecker;
 import gov.nist.csd.pm.pap.query.*;
