+++ src/test/java/gov/nist/csd/pm/pdp/modification/ProhibitionsModificationAdjudicatorTest.java
@@ -3,7 +3,6 @@ package gov.nist.csd.pm.pdp.modification;
 import gov.nist.csd.pm.pap.exception.PMException;
 import gov.nist.csd.pm.pap.graph.relationship.AccessRightSet;
 import gov.nist.csd.pm.pap.obligation.EventContext;
-import gov.nist.csd.pm.pap.op.PrivilegeChecker;
 import gov.nist.csd.pm.pap.prohibition.ContainerCondition;
 import gov.nist.csd.pm.pap.prohibition.ProhibitionSubject;
 import gov.nist.csd.pm.epp.EPP;
